feat(autoentrepreneur): ajoute la distribution des cotisations à la simulation AutoEntrepreneur
