+++ site/source/components/simulationExplanation/IndépendantExplanation.tsx
@@ -1,31 +1,20 @@
-import BarChartBranch from '@/components/BarChart'
-
 import '@/components/Distribution.css'
 
 import { DottedName } from 'modele-social'
-import { Trans, useTranslation } from 'react-i18next'
-import { useSelector } from 'react-redux'
+import { useTranslation } from 'react-i18next'
 import { useTheme } from 'styled-components'
 
-import RuleLink from '@/components/RuleLink'
+import { ÀQuoiServentMesCotisationsSection } from '@/components/simulationExplanation/ÀQuoiServentMesCotisationsSection'
+import { DroitsRetraite } from '@/components/simulationExplanation/DroitsRetraite'
 import StackedBarChart from '@/components/StackedBarChart'
-import { useEngine } from '@/components/utils/EngineContext'
-import { Message } from '@/design-system'
-import { Emoji } from '@/design-system/emoji'
-import { Strong } from '@/design-system/typography'
 import { H3 } from '@/design-system/typography/heading'
-import { Li, Ul } from '@/design-system/typography/list'
-import { SmallBody } from '@/design-system/typography/paragraphs'
-import { targetUnitSelector } from '@/store/selectors/simulationSelectors'
 
 import { Condition } from '../EngineValue/Condition'
-import Value from '../EngineValue/Value'
 import { WhenApplicable } from '../EngineValue/WhenApplicable'
 import { WhenNotApplicable } from '../EngineValue/WhenNotApplicable'
 import CotisationsForfaitaires from './IndépendantCotisationsForfaitaires'
 import CotisationsRégularisation from './IndépendantCotisationsRégularisation'
 import InstitutionsPartenaires from './InstitutionsPartenaires'
-import { DistributionSection } from './SalaryExplanation'
 
 export default function IndépendantExplanation() {
 	const { t } = useTranslation()
@@ -67,9 +56,7 @@ export default function IndépendantExplanation() {
 			</Condition>
 			<InstitutionsPartenaires />
 			<DroitsRetraite />
-			<DistributionSection>
-				<Distribution />
-			</DistributionSection>
+			<ÀQuoiServentMesCotisationsSection regroupement={CotisationsSection} />
 		</>
 	)
 }
@@ -99,145 +86,4 @@ const CotisationsSection: Partial<Record<DottedName, Array<string>>> = {
 	'protection sociale . formation': [
 		'dirigeant . indépendant . cotisations et contributions . formation professionnelle',
 	],
-}
-
-function Distribution() {
-	const targetUnit = useSelector(targetUnitSelector)
-	const engine = useEngine()
-	const distribution = (
-		Object.entries(CotisationsSection).map(([section, cotisations]) => [
-			section,
-			cotisations
-				.map((c) => engine.evaluate({ valeur: c, unité: targetUnit }))
-				.reduce(
-					(acc, evaluation) => acc + ((evaluation?.nodeValue as number) || 0),
-					0
-				),
-		]) as Array<[DottedName, number]>
-	)
-		.filter(([, value]) => value > 0)
-		.sort(([, a], [, b]) => b - a)
-
-	const maximum = Math.max(...distribution.map(([, value]) => value))
-
-	return (
-		<>
-			<div className="distribution-chart__container" role="list">
-				{distribution.map(([sectionName, value]) => (
-					<DistributionBranch
-						key={sectionName}
-						dottedName={sectionName}
-						value={value}
-						maximum={maximum}
-						role="listitem"
-					/>
-				))}
-			</div>
-		</>
-	)
-}
-
-type DistributionBranchProps = {
-	dottedName: DottedName
-	value: number
-	maximum: number
-	icon?: string
-	role?: string
-}
-
-function DistributionBranch({
-	dottedName,
-	value,
-	icon,
-	maximum,
-	...props
-}: DistributionBranchProps) {
-	const branche = useEngine().getRule(dottedName)
-
-	return (
-		<BarChartBranch
-			value={value}
-			maximum={maximum}
-			title={<RuleLink dottedName={dottedName} />}
-			icon={icon ?? branche.rawNode.icônes}
-			description={branche.rawNode.résumé}
-			unit="€"
-			{...props}
-		/>
-	)
-}
-
-function DroitsRetraite() {
-	const { t } = useTranslation()
-	const exonérationRetraiteActive = {
-		'une de ces conditions': [
-			'dirigeant . indépendant . cotisations et contributions . exonérations . ACRE',
-			'dirigeant . indépendant . cotisations et contributions . exonérations . pension invalidité',
-			'dirigeant . indépendant . PL . CNAVPL . exonération incapacité',
-			'dirigeant . indépendant . PL . CIPAV . exonération incapacité',
-		] as Array<DottedName>,
-	}
-
-	return (
-		<Trans i18nKey="pages.simulateurs.indépendant.retraite-droits-acquis">
-			<H3 as="h2">Retraite : droits acquis sur l'année</H3>
-			<Condition expression={exonérationRetraiteActive}>
-				<Message type="info" icon={<Emoji emoji="🚧" />} border={false}>
-					Le calcul des droits ouverts à la retraite n'est pas encore implémenté
-					pour les cas incluants des d'exonérations de cotisations (ACRE,
-					pension invalidité, etc).
-				</Message>
-			</Condition>
-			<Condition expression={{ '=': [exonérationRetraiteActive, 'non'] }}>
-				<Ul>
-					<Li>
-						Retraite de base :{' '}
-						<Value
-							expression="protection sociale . retraite . trimestres"
-							displayedUnit={t('trimestres acquis')}
-						/>
-					</Li>
-					<WhenApplicable dottedName="protection sociale . retraite . base . CNAVPL">
-						<Li>
-							Points de retraite de base acquis (CNAVPL) :{' '}
-							<Value
-								linkToRule
-								expression="protection sociale . retraite . base . CNAVPL"
-								displayedUnit={t('points')}
-							/>
-						</Li>
-					</WhenApplicable>
-					<WhenNotApplicable dottedName="protection sociale . retraite . base . CNAVPL">
-						<Li>
-							Revenu cotisé pris en compte pour la retraite de base :{' '}
-							<Value
-								linkToRule
-								unit="€/an"
-								expression="protection sociale . retraite . base"
-							/>
-						</Li>
-					</WhenNotApplicable>
-					<Li>
-						Points de retraite complémentaire acquis :{' '}
-						<WhenApplicable dottedName="protection sociale . retraite . complémentaire . RCI . points acquis">
-							<Value
-								expression="protection sociale . retraite . complémentaire . RCI . points acquis"
-								displayedUnit=""
-							/>{' '}
-							points acquis
-						</WhenApplicable>
-						<WhenNotApplicable dottedName="protection sociale . retraite . complémentaire . RCI . points acquis">
-							<Strong>non connue</Strong>
-							<WhenApplicable dottedName="dirigeant . indépendant . PL">
-								<SmallBody>
-									Ce simulateur ne gère pas les droits acquis de retraite
-									complémentaire pour les professions libérales
-								</SmallBody>
-							</WhenApplicable>
-						</WhenNotApplicable>
-					</Li>
-				</Ul>
-			</Condition>
-		</Trans>
-	)
 }
