+++ site/test/regressions/auto-entrepreneur.yaml
@@ -26,7 +26,7 @@ ACRE:
     entreprise . date de création: période . début d'année
     dirigeant . exonérations . ACRE: oui
   - dirigeant . auto-entrepreneur . revenu net: 30000 €/an
-    entreprise . date de création: 01/06/2023
+    entreprise . date de création: 01/06/2024
     dirigeant . exonérations . ACRE: oui
 
 activités:
